Added decrement and increment to counter cache columns on associated objects for ActiveRecord>=5.x

+++ lib/acts_as_paranoid/core.rb
@@ -117,7 +117,16 @@ module ActsAsParanoid
         run_callbacks :destroy do
           destroy_dependent_associations!
           # Handle composite keys, otherwise we would just use `self.class.primary_key.to_sym => self.id`.
-          self.class.delete_all!(Hash[[Array(self.class.primary_key), Array(self.id)].transpose]) if persisted?
+          if persisted?
+            affected_row_count = self.class.delete_all!(Hash[[Array(self.class.primary_key), Array(self.id)].transpose])
+
+            if ActiveRecord::VERSION::MAJOR >= 5
+              decrement_counters_on_associations affected_row_count
+            end
+            affected_row_count
+          else
+            true
+          end
           self.paranoid_value = self.class.delete_now_value
           freeze
         end
@@ -129,11 +138,18 @@ module ActsAsParanoid
         with_transaction_returning_status do
           run_callbacks :destroy do
             # Handle composite keys, otherwise we would just use `self.class.primary_key.to_sym => self.id`.
-            @_trigger_destroy_callback = if persisted?
-                                           self.class.delete_all(Hash[[Array(self.class.primary_key), Array(self.id)].transpose])
-                                         else
-                                           true
-                                         end
+            @_trigger_destroy_callback =
+              if persisted?
+                affected_row_count = self.class.delete_all(Hash[[Array(self.class.primary_key), Array(self.id)].transpose])
+
+                if ActiveRecord::VERSION::MAJOR >= 5
+                  decrement_counters_on_associations affected_row_count
+                end
+
+                affected_row_count
+              else
+                true
+              end
 
             self.paranoid_value = self.class.delete_now_value
             self
@@ -157,7 +173,7 @@ module ActsAsParanoid
       self.class.transaction do
         run_callbacks :recover do
           recover_dependent_associations(options[:recovery_window], options) if options[:recursive]
-
+          increment_counters_on_associations
           self.paranoid_value = self.class.paranoid_configuration[:recovery_value]
           self.save
         end
@@ -224,6 +240,34 @@ module ActsAsParanoid
 
     def paranoid_value=(value)
       self.send("#{self.class.paranoid_column}=", value)
+    end
+
+    def decrement_counters_on_associations affected_row_count
+      if affected_row_count > 0
+        each_counter_cached_association_reflection do |assoc_reflection|
+          foreign_key = assoc_reflection.foreign_key.to_sym
+          unless destroyed_by_association && destroyed_by_association.foreign_key.to_sym == foreign_key
+            associated_object = send(assoc_reflection.name)
+            counter_cache_column = assoc_reflection.counter_cache_column
+            associated_object.class.send(:decrement_counter, counter_cache_column, associated_object.id)
+          end
+        end
+      end
+    end
+
+    def increment_counters_on_associations
+      each_counter_cached_association_reflection do |assoc_reflection|
+        foreign_key = assoc_reflection.foreign_key.to_sym
+        associated_object = send(assoc_reflection.name)
+        counter_cache_column = assoc_reflection.counter_cache_column
+        associated_object.class.send(:increment_counter, counter_cache_column, associated_object.id)
+      end
+    end
+
+    def each_counter_cached_association_reflection
+      _reflections.each do |name, reflection|
+        yield reflection if reflection.belongs_to? && reflection.counter_cache_column == "#{self.class.name.underscore.pluralize}_count"
+      end
     end
   end
 end
